redirect to current year page
